requirements.txt
asgiref
Django
djangorestframework
Pillow
psycopg2-binary
python-magic-bin
pytz
sqlparse
